app/Engine/Launchpad/OPENAI_CONNECTOR.py
import openai
from dotenv import load_dotenv
import os

load_dotenv()
openai.api_key = os.getenv("OPENAI_API_KEY")



def getResponse(prompt, temperature, max_tokens, top_p, frequency_penalty, presence_penalty):
    response = openai.Completion.create(
        model="text-davinci-003",
        prompt=prompt,
        temperature=temperature,
        max_tokens=max_tokens,
        top_p=top_p,
        frequency_penalty=frequency_penalty,
        presence_penalty=presence_penalty,
    )

    response_text = response.choices[0].text
    return response_text