requirements.txt
django==4.2.1
openai==0.27.8
python-dotenv==1.0.0
